quic: move connection_id_generator default extension to subdirectory
This makes it follow the Envoy extension pattern of
<category>/<extension-name>/files.
