+++ source/extensions/quic/connection_id_generator/deterministic/envoy_deterministic_connection_id_generator.h
@@ -10,6 +10,9 @@
 
 namespace Envoy {
 namespace Quic {
+namespace Extensions {
+namespace ConnectionIdGenerator {
+namespace Deterministic {
 
 // This class modifies connection ids that are too long in an Envoy fashion.
 class EnvoyDeterministicConnectionIdGenerator : public quic::DeterministicConnectionIdGenerator {
@@ -44,5 +47,8 @@ private:
 #endif
 };
 
+} // namespace Deterministic
+} // namespace ConnectionIdGenerator
+} // namespace Extensions
 } // namespace Quic
 } // namespace Envoy
