+++ source/extensions/quic/connection_id_generator/deterministic/envoy_deterministic_connection_id_generator_config.cc
@@ -1,11 +1,14 @@
-#include "source/extensions/quic/connection_id_generator/envoy_deterministic_connection_id_generator_config.h"
+#include "source/extensions/quic/connection_id_generator/deterministic/envoy_deterministic_connection_id_generator_config.h"
 
 #include "envoy/extensions/quic/connection_id_generator/v3/envoy_deterministic_connection_id_generator.pb.h"
 
-#include "source/extensions/quic/connection_id_generator/envoy_deterministic_connection_id_generator.h"
+#include "source/extensions/quic/connection_id_generator/deterministic/envoy_deterministic_connection_id_generator.h"
 
 namespace Envoy {
 namespace Quic {
+namespace Extensions {
+namespace ConnectionIdGenerator {
+namespace Deterministic {
 
 ProtobufTypes::MessagePtr
 EnvoyDeterministicConnectionIdGeneratorConfigFactory::createEmptyConfigProto() {
@@ -22,5 +25,8 @@ EnvoyDeterministicConnectionIdGeneratorConfigFactory::createQuicConnectionIdGene
 REGISTER_FACTORY(EnvoyDeterministicConnectionIdGeneratorConfigFactory,
                  EnvoyQuicConnectionIdGeneratorConfigFactory);
 
+} // namespace Deterministic
+} // namespace ConnectionIdGenerator
+} // namespace Extensions
 } // namespace Quic
 } // namespace Envoy
