+++ source/extensions/quic/connection_id_generator/deterministic/envoy_deterministic_connection_id_generator_config.h
@@ -6,18 +6,24 @@
 
 namespace Envoy {
 namespace Quic {
+namespace Extensions {
+namespace ConnectionIdGenerator {
+namespace Deterministic {
 
 class EnvoyDeterministicConnectionIdGeneratorConfigFactory
-    : public EnvoyQuicConnectionIdGeneratorConfigFactory {
+    : public Quic::EnvoyQuicConnectionIdGeneratorConfigFactory {
 public:
   // EnvoyQuicConnectionIdGeneratorConfigFactory.
   ProtobufTypes::MessagePtr createEmptyConfigProto() override;
-  EnvoyQuicConnectionIdGeneratorFactoryPtr
+  Quic::EnvoyQuicConnectionIdGeneratorFactoryPtr
   createQuicConnectionIdGeneratorFactory(const Protobuf::Message& config) override;
   std::string name() const override { return "envoy.quic.deterministic_connection_id_generator"; }
 };
 
 DECLARE_FACTORY(EnvoyDeterministicConnectionIdGeneratorConfigFactory);
 
+} // namespace Deterministic
+} // namespace ConnectionIdGenerator
+} // namespace Extensions
 } // namespace Quic
 } // namespace Envoy
